\name{battese}
\alias{battese}
\docType{data}
\title{Battese Fuller data}
\description{
  
}
\usage{data(battese)}
\format{
 This table contains the data of the paper  "An Error-Components Model for Prediction of County Crop Areas Using Survey and Satellite Data" by 
   George E. Battese, Rachel M. Harter and Wayne A. Fuller (Vol. 83, No. 401 (Mar., 1988), pp. 28-36)
  \describe{
    \name{battese}
    \item{county }{County identifier} 
    \item{countyn}{Number of segments in county}
    \item{crnsrvy}{Reported hectares of corn}
    \item{soysrvy}{Reported hectares of soybeans}
    \item{crnpxel}{Number of pixels of corn in sample segments}
    \item{soypxel}{Number of pixels of Soybean in sample segments}
    \item{crnmean}{Mean number of pixels of corn per segment }
    \item{soymean}{Mean number of pixels of soybean per segment}
  }
}
\source{
  An Error-Components Model for Prediction of County Crop Areas Using Survey and Satellite Data by 
   George E. Battese, Rachel M. Harter and Wayne A. Fuller (Vol. 83, No. 401 (Mar., 1988), pp. 28-36)}
\examples{
  data(battese)
}
\keyword{datasets}
  